requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
pydantic
httpx
google-generativeai
rasterio
shapely
pandas
openpyxl
python-docx
sentence-transformers
numpy
scikit-learn
pypdf
geopandas
